requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.9.0
dnspython==2.4.2
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.108.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.0
icecream==2.1.3
idna==3.6
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
langchain-openai==0.0.2
langserve==0.0.38
langsmith==0.0.77
loguru==0.7.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.6.1
orjson==3.9.10
packaging==23.2
pinecone-client==2.2.4
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sse-starlette==1.8.2
starlette==0.32.0.post1
tenacity==8.2.3
tiktoken==0.5.2
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
win32-setctime==1.1.0
yarl==1.9.4
